Rename new util to getProductErrorMessage
