+++ packages/js/product-editor/src/components/header/hooks/use-preview/use-preview.tsx
@@ -13,7 +13,7 @@ import { MouseEvent } from 'react';
  * Internal dependencies
  */
 import { useValidations } from '../../../../contexts/validation-context';
-import { WPError } from '../../../../utils/get-error-message';
+import { WPError } from '../../../../utils/get-product-error-message';
 
 export function usePreview( {
 	disabled,
