+++ packages/js/product-editor/src/components/header/hooks/use-publish/use-publish.tsx
@@ -12,7 +12,7 @@ import { MouseEvent } from 'react';
  * Internal dependencies
  */
 import { useValidations } from '../../../../contexts/validation-context';
-import { WPError } from '../../../../utils/get-error-message';
+import { WPError } from '../../../../utils/get-product-error-message';
 
 export function usePublish( {
 	disabled,
